devided into components. doesnt work. need props
